Move server-side API tests to server directory.
